検索結果画面をコンポーネント分割した
