+++ app/javascript/packs/app.vue
@@ -12,37 +12,10 @@
       />
       <button @click="submitSearch">search</button>
     </div>
-    <h2>results</h2>
-    <div v-show="state.results.length !== 0">
-      <table>
-        <thead>
-          <tr>
-            <th>楽曲名</th>
-            <th>アーティスト名</th>
-            <th>Spotify</th>
-            <th>Apple Music</th>
-            <th>KKBOX</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody v-for="result in state.results" :key="result.isrc">
-          <tr>
-            <td>{{ result.title }}</td>
-            <td>{{ result.artist }}</td>
-            <td>{{ result.spotifyUrl }}</td>
-            <td>{{ result.appleMusicUrl }}</td>
-            <td>{{ result.kkboxUrl }}</td>
-            <td>
-              <button @click="copyUrlsToClipBoard(result)">
-                この曲をシェアする
-              </button>
-              <div v-show="state.isCopiedUrls">クリップボードにURLがコピーされました</div>
-              <div v-show="state.isFailedUrls">URLのコピーが失敗しました</div>
-            </td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
+    <results-list
+      v-show="state.results.length !== 0"
+      :results="state.results"
+    ></results-list>
   </div>
 </template>
 
@@ -50,17 +23,17 @@
 import { defineComponent, reactive } from "vue";
 import axios from "axios";
 import { parseResponseData } from "./utils/parseResponseData";
-
-const TOOLTIP_SHOW_TIME = 3000;
+import resultsList from "./components/resultsList.vue";
 
 export default defineComponent({
   name: "App",
+  components: {
+    "results-list": resultsList,
+  },
   setup() {
     const state = reactive({
       keywords: "",
       results: [],
-      isCopiedUrls: false,
-      isFailedUrls: false,
     });
 
     async function submitSearch() {
@@ -75,43 +48,9 @@ export default defineComponent({
       }
     }
 
-    function copyUrlsToClipBoard(result) {
-      navigator.clipboard
-        .writeText(urlsText(result))
-        .then(() => {
-          state.isCopiedUrls = true;
-          setTimeout(function () {
-            state.isCopiedUrls = false;
-          }, TOOLTIP_SHOW_TIME);
-        })
-        .catch(() => {
-          state.isFailedCopyUrls = true;
-          setTimeout(function () {
-            state.isFailedCopyUrls = false;
-          }, TOOLTIP_SHOW_TIME);
-        });
-    }
-
-    function urlsText(result) {
-      let urls = [`${result.title}(${result.artist})`];
-
-      if (result.spotifyUrl) {
-        urls.push(`Spotify\n${result.spotifyUrl}`);
-      }
-      if (result.appleMusicUrl) {
-        urls.push(`Apple Music\n${result.appleMusicUrl}`);
-      }
-      if (result.kkboxUrl) {
-        urls.push(`KKBOX\n${result.kkboxUrl}`);
-      }
-
-      return urls.join("\n");
-    }
-
     return {
       state,
       submitSearch,
-      copyUrlsToClipBoard,
     };
   },
 });
