+++ app/javascript/packs/components/resultsList.vue
@@ -1,0 +1,42 @@
+<template>
+  <div>
+    <h2>results</h2>
+    <div v-show="results.length !== 0">
+      <table>
+        <thead>
+          <tr>
+            <th>楽曲名</th>
+            <th>アーティスト名</th>
+            <th>Spotify</th>
+            <th>Apple Music</th>
+            <th>KKBOX</th>
+            <th></th>
+          </tr>
+        </thead>
+        <tbody v-for="result in results" :key="result.isrc">
+          <result :result="result"></result>
+        </tbody>
+      </table>
+    </div>
+  </div>
+</template>
+
+<script lang="ts">
+import { defineComponent } from "vue";
+import result from "./result";
+
+export default defineComponent({
+  name: "ResultList",
+  components: {
+    result: result,
+  },
+  props: {
+    results: {
+      type: Array,
+      default: () => [],
+    },
+  },
+});
+</script>
+
+<style scoped></style>
