+++ app/javascript/packs/types/responseData.ts
@@ -1,5 +1,4 @@
 export type responseData = {
-  // result_struct = Struct.new(:isrc, :title, :artist, :spotify_url, :apple_music_url, :kkbox_url, keyword_init: true)
   isrc: string;
   title: string;
   artist: string;
