+++ app/javascript/packs/types/result.ts
@@ -1,0 +1,8 @@
+export type result = {
+  isrc: string;
+  title: string;
+  artist: string;
+  spotifyUrl: string;
+  appleMusicUrl: string;
+  kkboxUrl: string;
+};
